Move/rename MFA data associator
Also minor flake8 fixes included, and change math.inf to np.inf
